Add repler_hostname to nginx.conf.template
At the moment the repler service is repeatedly going up and down.
This is not caused by commits to the repler repo, but instead
appears to be caused by the container failing its health checks.
This is causing a lot of noise in the Kosli snapshots UX for the
two cyber-dojo Environments, aws-beta and aws-prod, which we
would like to get rid of.